refactor: extract shared helpers
