+++ tally-list-card-editor.js
@@ -1,4 +1,5 @@
 import { LitElement, html, css } from 'https://unpkg.com/lit?module';
+import { translate, fireEvent } from './tally-list-utils.js';
 const CARD_VERSION = '09.08.2025';
 
 const TL_STRINGS = {
@@ -54,26 +55,8 @@ const TL_STRINGS = {
   },
 };
 
-function detectLang(hass, override = 'auto') {
-  if (override && override !== 'auto') return override;
-  const lang =
-    hass?.language || hass?.locale?.language || navigator.language || 'en';
-  return lang.toLowerCase().startsWith('de') ? 'de' : 'en';
-}
-
 function t(hass, override, key) {
-  const lang = detectLang(hass, override);
-  return TL_STRINGS[lang][key] || TL_STRINGS.en[key] || key;
-}
-
-function fireEvent(node, type, detail = {}, options = {}) {
-  node.dispatchEvent(
-    new CustomEvent(type, {
-      detail,
-      bubbles: options.bubbles ?? true,
-      composed: options.composed ?? true,
-    })
-  );
+  return translate(hass, override, TL_STRINGS, key);
 }
 
 class TallyListCardEditor extends LitElement {
