+++ tally-list-card.js
@@ -1,6 +1,7 @@
 // Tally List Card
 import { LitElement, html, css } from 'https://unpkg.com/lit?module';
 import { repeat } from 'https://unpkg.com/lit/directives/repeat.js?module';
+import { detectLang, translate, fireEvent } from './tally-list-utils.js';
 const CARD_VERSION = '09.08.2025';
 
 const TL_STRINGS = {
@@ -127,26 +128,8 @@ const TL_STRINGS = {
   },
 };
 
-function detectLang(hass, override = 'auto') {
-  if (override && override !== 'auto') return override;
-  const lang =
-    hass?.language || hass?.locale?.language || navigator.language || 'en';
-  return lang.toLowerCase().startsWith('de') ? 'de' : 'en';
-}
-
 function t(hass, override, key) {
-  const lang = detectLang(hass, override);
-  return TL_STRINGS[lang][key] || TL_STRINGS.en[key] || key;
-}
-
-function fireEvent(node, type, detail = {}, options = {}) {
-  node.dispatchEvent(
-    new CustomEvent(type, {
-      detail,
-      bubbles: options.bubbles ?? true,
-      composed: options.composed ?? true,
-    })
-  );
+  return translate(hass, override, TL_STRINGS, key);
 }
 
 function relevantStatesChanged(newHass, oldHass, entities) {
@@ -157,9 +140,7 @@ function relevantStatesChanged(newHass, oldHass, entities) {
   return false;
 }
 
-const navLang = (navigator.language || '').toLowerCase().startsWith('de')
-  ? 'de'
-  : 'en';
+const navLang = detectLang();
 window.customCards = window.customCards || [];
 window.customCards.push({
   type: 'tally-list-card',
